+++ pagerduty/api_client.py
@@ -6,6 +6,7 @@ import time
 from copy import deepcopy
 from random import random
 from typing import Optional, Union
+from warnings import warn
 
 # PyPI
 from requests import Response, Session
@@ -13,17 +14,16 @@ from requests import __version__ as REQUESTS_VERSION
 from requests.exceptions import RequestException
 from urllib3.exceptions import PoolError
 from urllib3.exceptions import HTTPError as Urllib3HttpError
- 
+
 # Local
+from . auth_method import AuthMethod
 from . version import __version__
 from . errors import (
     Error,
-    HttpError,
-    UrlError
+    HttpError
 )
 from . common import (
     TIMEOUT,
-    last_4,
     normalize_url
 )
 
@@ -45,8 +45,9 @@ class ApiClient(Session):
       :attr:`permitted_methods` list, and will raise :class:`Error` for
       any other HTTP methods.
 
-    :param api_key:
-        The API secret to use for authentication in HTTP requests
+    :param auth_method:
+        The authentication method to use for API requests, should be an instance
+        of the AuthMethod class.
     :param debug:
         Sets :attr:`print_debug`. Set to ``True`` to enable verbose command line
         output.
@@ -127,44 +128,49 @@ class ApiClient(Session):
     """
 
     url = ""
+    """
+    The base URL for the API being called (usually https://api.pagerduty.com, but
+    this can vary depending on the specific API being accessed).
+    """
 
-    def __init__(self, api_key: str, debug=False):
+    def __init__(self, auth_method: AuthMethod, debug=False):
         self.parent = super(ApiClient, self)
         self.parent.__init__()
-        self.api_key = api_key
+        self.auth_method = auth_method
         self.log = logging.getLogger(__name__)
         self.print_debug = debug
         self.retry = {}
 
-    def after_set_api_key(self):
+    def after_set_auth_method(self):
         """
-        Setter hook for setting or updating the API key.
-
+        Setter hook for setting or updating the authentication method.
         Child classes should implement this to perform additional steps.
         """
         pass
 
     @property
-    def api_key(self) -> str:
+    def auth_method(self) -> AuthMethod:
         """
-        Property representing the credential used for accessing the given API.
+        Property representing the authentication method used for API requests.
         """
-        return self._api_key
+        return self._auth_method
 
-    @api_key.setter
-    def api_key(self, api_key: str):
-        if not (isinstance(api_key, str) and api_key):
-            raise ValueError("API credential must be a non-empty string.")
-        self._api_key = api_key
-        self.headers.update(self.auth_header)
-        self.after_set_api_key()
+    @auth_method.setter
+    def auth_method(self, auth_method: AuthMethod):
+        if not (isinstance(auth_method, AuthMethod)):
+            raise ValueError("auth_method must be an instance of the AuthMethod class")
+
+        self._auth_method = auth_method
+        self.after_set_auth_method()
 
     @property
     def auth_header(self) -> dict:
         """
-        Generates the header with the API credential used for authentication.
+        Generates the Authorization header based on auth_method provided.
         """
-        raise NotImplementedError
+        warn("Property ApiClient.auth_header is deprecated. " +
+            "Use ApiClient.auth_method.auth_header instead.")
+        return self.auth_method.auth_header
 
     def cooldown_factor(self) -> float:
         return self.sleep_timer_base*(1+self.stagger_cooldown*random())
@@ -202,14 +208,19 @@ class ApiClient(Session):
         :returns:
             The final list of headers to use in the request
         """
+        # Utilize any defaults that the implementer has set via the upstream interface:
         headers = deepcopy(self.headers)
+        # Override the default user-agent with the per-class user_agent property:
         headers['User-Agent'] = self.user_agent
         # A universal convention: whenever sending a POST, PUT or PATCH, the
-        # Content-Type header is "application/json".
+        # Content-Type header must be "application/json":
         if method in ('POST', 'PUT', 'PATCH'):
             headers['Content-Type'] = 'application/json'
+        # Add headers passed in per-request as an additional argument:
         if type(user_headers) is dict:
             headers.update(user_headers)
+        # Add authentication header, if the auth_method defines it:
+        headers.update(self.auth_method.auth_header)
         return headers
 
     @property
@@ -279,7 +290,14 @@ class ApiClient(Session):
             'timeout': self.timeout
         })
 
-        # Special changes to user-supplied parameters, for convenience
+        # Add authentication parameter, if the API requires it and it is a request type
+        # that includes a body:
+        if method in ('POST', 'PUT', 'PATCH'):
+            for body_key in ('json', 'data'):
+                if body_key in req_kw and type(req_kw[body_key]) is dict:
+                    req_kw[body_key].update(self.auth_method.auth_param)
+
+        # Special changes to user-supplied query parameters, for convenience:
         if 'params' in kwargs and kwargs['params']:
             req_kw['params'] = self.normalize_params(kwargs['params'])
 
@@ -386,7 +404,7 @@ class ApiClient(Session):
     @property
     def trunc_key(self) -> str:
         """Truncated key for secure display/identification purposes."""
-        return last_4(self.api_key)
+        return self.auth_method.trunc_secret
 
     @property
     def user_agent(self) -> str: