+++ pagerduty/oauth_token_client.py
@@ -9,14 +9,40 @@ from copy import deepcopy
 from requests import Response
 
 from . api_client import ApiClient
+from . auth_method import BodyParameterAuthMethod
 from . common import (
     datetime_to_relative_seconds,
     relative_seconds_to_datetime,
     successful_response,
-    try_decoding
+    try_decoding,
+    last_4
 )
 from . errors import ServerHttpError
 from . rest_api_v2_client import RestApiV2Client
+
+class ClientCredentialsAuthMethod(BodyParameterAuthMethod):
+    """
+    AuthMethod used in OAuth token exchange requests.
+
+    This AuthMethod requires the client secret and client ID, which are then transmitted
+    to identity.pagerduty.com as parameters in the HTML form-encoded body.
+
+    :param client_secret:
+        The client secret used for creating OAuth tokens for the application
+    :param client_id:
+        The client ID of the provisioned OAuth application
+    """
+
+    def __init__(self, client_secret, client_id):
+        self.secret = client_secret
+        self.client_id = client_id # Additional required parameter
+
+    @property
+    def auth_param(self) -> dict:
+        return {
+            "client_id": self.client_id,
+            "client_secret": self.secret
+        }
 
 class OAuthTokenClient(ApiClient):
     """
@@ -73,8 +99,9 @@ class OAuthTokenClient(ApiClient):
         """
         Create an OAuth token client
         """
-        super(OAuthTokenClient, self).__init__(client_secret, debug=debug)
-        self.client_id = client_id
+        auth_method = ClientCredentialsAuthMethod(client_secret, client_id)
+
+        super(OAuthTokenClient, self).__init__(auth_method, debug=debug)
 
     def amended_auth_response(self, response: Response) -> dict:
         """
@@ -98,10 +125,6 @@ class OAuthTokenClient(ApiClient):
         })
         return response_json
 
-    @property
-    def auth_header(self) -> dict:
-        return {}
-
     def authorize_url(self, scope: str, redirect_uri: str) -> str:
         """
         The authorize URL in PagerDuty that the end user will visit to authorize the app
@@ -115,20 +138,10 @@ class OAuthTokenClient(ApiClient):
             The formatted authorize URL.
         """
         return self.get_authorize_url(
-            self.client_id,
+            self.auth_method.client_id,
             scope,
             redirect_uri
         )
-
-    @property
-    def api_key(self):
-        return self._api_key
-
-    @api_key.setter
-    def api_key(self, api_key: str):
-        if not (isinstance(api_key, str) and api_key):
-            raise ValueError("Client secret must be a non-empty string.")
-        self._api_key = api_key
 
     @classmethod
     def get_authorize_url(cls, client_id: str, scope: str, redirect_uri: str) -> str:
@@ -171,21 +184,19 @@ class OAuthTokenClient(ApiClient):
         * :attr:`get_scoped_app_token`
         * :attr:`get_new_token_from_code_with_pkce`
 
+        :param kw:
+            Keyword parameters passed to this method are interpreted as parameters to
+            set in the body of the request.
         :returns:
             The JSON response from ``identity.pagerduty.com`` as a dictionary after
             amending via :attr:`amended_auth_response`. It should contain a key
             ``access_token`` with the new token and ``expiration_date`` containing the
             date and time when the token will expire in ISO8601 format.
         """
-        params = {
-            "client_id": self.client_id,
-            "client_secret": self.api_key
-        }
-        params.update(kw)
         return self.amended_auth_response(self.post(
             '/oauth/token',
-            data = params,
-            headers = {
+            data = deepcopy(kw),
+            headers = { # Overrides the default which would be `application/json`:
                 "Content-Type": "application/x-www-form-urlencoded"
             }
         ))