+++ CHANGELOG.rst
@@ -1,3 +1,13 @@
+**2025-09-24: API authentication interface refactor - Version 5.0.0**
+
+In this version, the details of how authentication are performed are offloaded to a new class, ``AuthMethod``. Each instance of ``ApiClient`` have a new property, ``auth_method``, which is an instance of ``AuthMethod``. This provides a generic framework for implementing all the forms of authentication (using headers or parameters in the request body) utilized in all of PagerDuty's APIs, minimizing redundant code in ``ApiClient`` subclasses.
+
+The primary motivation of this change was to make the behavior of REST API v2 client objects more stable and predictable when swapping out their API credentials mid-process. This is needed for the purposes of an internal project at PagerDuty.
+
+* **Breaking Changes:**
+   - The ``api_key`` property has been removed from all client classes except for ``pagerduty.RestApiV2BaseClient`` and subclasses thereof, wherein it is deprecated.
+   - The hook method ``after_set_api_key`` will be ignored in all classes where the ``api_key`` property has been removed.
+
 **2025-09-08: Bugfix: iter_all parameter serialization - Version 4.1.1**
 
 * Fixes a regression wherein a boolean is serialized to the capitalized ``True`` / ``False`` for the ``total`` parameter in classic pagination, whereas the API requires lower case values.